Create NIME 2021 Workshop
Copied the pub pub md file across. Since this is the proposal file I don't expect this to be the front-facing workshop static page that we will have in the end. We should change the format a little bit. For now we should expand on the discussion topics, see lines: 205-227